Moving mongo package into db package
