+++ src/app/beer_garden/api/http/authorization.py
@@ -11,7 +11,7 @@ from passlib.apps import custom_app_context
 from tornado.web import HTTPError
 
 import beer_garden.api.http
-from beer_garden.bg_utils.mongo.models import Principal, Role
+from beer_garden.db.mongo.models import Principal, Role
 
 
 class Permissions(Enum):
