+++ src/app/beer_garden/api/http/handlers/v2/request.py
@@ -1,7 +1,7 @@
 import json
 
 import beer_garden.bg_utils
-from beer_garden.bg_utils.mongo.models import Request
+from beer_garden.db.mongo.models import Request
 from beer_garden.api.http.authorization import authenticated, Permissions
 from beer_garden.api.http.base_handler import BaseHandler
 from beer_garden.api.http.client import ExecutorClient
