+++ src/app/beer_garden/db/mongo/models.py
