+++ src/app/beer_garden/metrics.py
@@ -8,7 +8,7 @@ from prometheus_client import Gauge, Counter, Summary
 from prometheus_client.exposition import MetricsHandler
 from prometheus_client.registry import REGISTRY
 
-from beer_garden.bg_utils.mongo.models import Request
+from beer_garden.db.mongo.models import Request
 
 
 class PrometheusServer(StoppableThread):
