+++ src/app/beer_garden/mongo_pruner.py
@@ -6,7 +6,7 @@ from typing import Tuple, List
 from brewtils.stoppable_thread import StoppableThread
 from mongoengine import Q
 
-from beer_garden.bg_utils.mongo.models import Event, Request
+from beer_garden.db.mongo.models import Event, Request
 
 
 class MongoPruner(StoppableThread):
