+++ src/app/beer_garden/requests.py
@@ -18,8 +18,8 @@ from mongoengine import Q
 from requests import Session
 
 import beer_garden
-from beer_garden.bg_utils.mongo.models import Choices, Request, System
-from beer_garden.bg_utils.mongo.parser import MongoParser
+from beer_garden.db.mongo.models import Choices, Request, System
+from beer_garden.db.mongo.parser import MongoParser
 from beer_garden.events import publish_event
 from beer_garden.metrics import request_created, request_started, request_completed
