+++ src/app/test/api/http/integration/controllers/request_api_test.py
@@ -2,8 +2,8 @@ import unittest
 
 from mock import Mock, patch
 
-from beer_garden.bg_utils.mongo.models import Request
-from beer_garden.bg_utils.mongo.parser import MongoParser
+from beer_garden.db.mongo.models import Request
+from beer_garden.db.mongo.parser import MongoParser
 
 
 @unittest.skip("TODO")
