+++ src/app/test/api/http/unit/handlers/v1/job_test.py
@@ -6,7 +6,7 @@ import unittest
 from datetime import datetime
 from mock import patch
 
-from beer_garden.bg_utils.mongo.models import Job, DateTrigger, RequestTemplate
+from beer_garden.db.mongo.models import Job, DateTrigger, RequestTemplate
 from .. import TestHandlerBase
