+++ src/app/test/api/http/unit/handlers/v1/roles_test.py
@@ -2,7 +2,7 @@
 import pytest
 from tornado.httpclient import HTTPRequest
 
-from beer_garden.bg_utils.mongo.models import Role
+from beer_garden.db.mongo.models import Role
 from brewtils.models import PatchOperation
 from brewtils.schema_parser import SchemaParser
