+++ src/app/test/api/http/unit/handlers/v1/users_test.py
@@ -4,7 +4,7 @@ import json
 import pytest
 from tornado.httpclient import HTTPRequest
 
-from beer_garden.bg_utils.mongo.models import Principal, Role
+from beer_garden.db.mongo.models import Principal, Role
 from brewtils.models import PatchOperation
 from brewtils.schema_parser import SchemaParser
