+++ src/app/test/mongo_pruner_test.py
@@ -4,7 +4,7 @@ from datetime import timedelta
 import pytest
 from mock import MagicMock, Mock, patch
 
-from beer_garden.bg_utils.mongo.models import Request, Event
+from beer_garden.db.mongo.models import Request, Event
 from beer_garden.mongo_pruner import MongoPruner
